fix(loans): show recognition settings when editing a loan/mortgage
The overpayment recognition settings (and the interest category) were placed
inside MortgageFields' create-only (`!isEditing`) payment block, so on the
edit form -- the very place the gear menu used to cover -- none of them showed.
An existing loan could no longer be configured at all.

Fold the interest category into the shared OverpaymentRecognitionFields and
render it unconditionally in both LoanFields and MortgageFields, so interest
category + overpayment category / payee / memo are available on create and
edit. Only the genuine payment-setup fields (frequency, first date, payment
account) stay create-only. Updated the affected tests.

Accounts suite (596) green; tsc + ESLint clean.
