+++ frontend/src/components/accounts/MortgageFields.tsx
@@ -1,15 +1,13 @@
 'use client';
 
-import { useState, useEffect, useCallback, useMemo } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { useTranslations } from 'next-intl';
 import { UseFormRegister, UseFormSetValue, FieldErrors } from 'react-hook-form';
 import { Input } from '@/components/ui/Input';
 import { DateInput } from '@/components/ui/DateInput';
 import { Select } from '@/components/ui/Select';
-import { Combobox } from '@/components/ui/Combobox';
 import { Account, MortgageAmortizationPreview, MortgagePaymentFrequency } from '@/types/account';
 import { Category } from '@/types/category';
-import { buildCategoryTree } from '@/lib/categoryUtils';
 import { accountsApi } from '@/lib/accounts';
 import { OverpaymentRecognitionFields } from './OverpaymentRecognitionFields';
 import { buildAccountDropdownOptions } from '@/lib/account-utils';
@@ -205,25 +203,6 @@ export function MortgageFields({
     return () => clearTimeout(timer);
   }, [calculateMortgagePreview]);
 
-  const interestCategoryOptions = useMemo(() =>
-    buildCategoryTree(categories).map(({ category }) => {
-      const parentCategory = category.parentId
-        ? categories.find(c => c.id === category.parentId)
-        : null;
-      return {
-        value: category.id,
-        label: parentCategory ? `${parentCategory.name}: ${category.name}` : category.name,
-      };
-    }),
-  [categories]);
-
-  const initialInterestCategoryName = useMemo(() => {
-    if (!selectedInterestCategoryId) return '';
-    const cat = categories.find(c => c.id === selectedInterestCategoryId);
-    if (!cat) return '';
-    const parent = cat.parentId ? categories.find(c => c.id === cat.parentId) : null;
-    return parent ? `${parent.name}: ${cat.name}` : cat.name;
-  }, [selectedInterestCategoryId, categories]);
 
   return (
     <div className="space-y-4 p-4 bg-purple-50 dark:bg-purple-900/20 rounded-lg border border-purple-200 dark:border-purple-800">
@@ -367,40 +346,18 @@ export function MortgageFields({
             />
           </div>
 
-          <div className="grid grid-cols-2 gap-4">
-            <Select
-              label={t('mortgageFields.paymentFromAccount')}
-              options={[
-                { value: '', label: t('mortgageFields.selectAccount') },
-                ...buildAccountDropdownOptions(
-                  accounts,
-                  () => true,
-                  (a) => `${a.name} (${a.currencyCode})`,
-                ),
-              ]}
-              error={errors.sourceAccountId?.message as string | undefined}
-              {...register('sourceAccountId')}
-            />
-
-            <Combobox
-              label={t('mortgageFields.interestCategory')}
-              placeholder={t('mortgageFields.selectCategory')}
-              options={interestCategoryOptions}
-              value={selectedInterestCategoryId}
-              initialDisplayValue={initialInterestCategoryName}
-              onChange={handleInterestCategoryChange}
-              error={errors.interestCategoryId?.message as string | undefined}
-            />
-          </div>
-
-          <OverpaymentRecognitionFields
-            categories={categories}
-            selectedOverpaymentCategoryId={selectedOverpaymentCategoryId}
-            onOverpaymentCategoryChange={handleOverpaymentCategoryChange}
-            selectedOverpaymentPayeeId={selectedOverpaymentPayeeId}
-            onOverpaymentPayeeChange={handleOverpaymentPayeeChange}
-            register={register}
-            errors={errors}
+          <Select
+            label={t('mortgageFields.paymentFromAccount')}
+            options={[
+              { value: '', label: t('mortgageFields.selectAccount') },
+              ...buildAccountDropdownOptions(
+                accounts,
+                () => true,
+                (a) => `${a.name} (${a.currencyCode})`,
+              ),
+            ]}
+            error={errors.sourceAccountId?.message as string | undefined}
+            {...register('sourceAccountId')}
           />
 
           {/* Mortgage Amortization Preview */}
@@ -456,6 +413,20 @@ export function MortgageFields({
           )}
         </>
       )}
+
+      {/* Payment recognition (interest category + overpayment category / payee /
+          memo). Always shown so an existing loan can be configured on edit. */}
+      <OverpaymentRecognitionFields
+        categories={categories}
+        selectedInterestCategoryId={selectedInterestCategoryId}
+        onInterestCategoryChange={handleInterestCategoryChange}
+        selectedOverpaymentCategoryId={selectedOverpaymentCategoryId}
+        onOverpaymentCategoryChange={handleOverpaymentCategoryChange}
+        selectedOverpaymentPayeeId={selectedOverpaymentPayeeId}
+        onOverpaymentPayeeChange={handleOverpaymentPayeeChange}
+        register={register}
+        errors={errors}
+      />
     </div>
   );
 }
